game.rbを作成してmain.rbからGameクラスを切り離しました
